% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unittest.R
\name{unittest}
\alias{unittest}
\title{Run unit tests}
\usage{
unittest(ns, standalone = TRUE, reporter = "progress")
}
\arguments{
\item{ns}{namespace (package) to test}

\item{standalone}{if \code{TRUE}, appropriate \link[testthat:Reporter]{testthat::Reporter}s are used, if
\code{FALSE} the exectution assumes reporters are started already.}

\item{reporter}{\link[testthat:Reporter]{testthat::Reporter} to be used.}
}
\value{
Invisibly returns a data.frame with the test results
}
\description{
If \link[testthat:testthat-package]{testthat::testthat-package} is available, run the unit tests and display
the results.
}
\examples{
unittest("hySpc.testthat")
}
\author{
Claudia Beleites
}
\keyword{programming}
\keyword{utilities}
